+++ Annex/src/utopia/annex/model/request/PostSpiritRequest.scala
@@ -89,6 +89,7 @@ trait PostSpiritRequest[+S <: Spirit with ModelConvertible, +A] extends ApiReque
 	
 	def method: Method = Post
 	override def path = spirit.postPath
+	override def pathParams: Model = Model.empty
 	override def body = Left(spirit.postBody)
 	
 	override def deprecated = false