+++ Annex/test/utopia/annex/test/GetEcho.scala
@@ -7,8 +7,8 @@ import utopia.annex.controller.ApiClient
 import utopia.annex.model.request.ApiRequest
 import utopia.annex.model.response.RequestResult
 import utopia.disciple.model.request.Body
-import utopia.flow.generic.model.immutable.{Model, Value}
 import utopia.flow.generic.casting.ValueConversions._
+import utopia.flow.generic.model.immutable.{Model, Value}
 
 import scala.concurrent.Future
 
@@ -17,7 +17,8 @@ import scala.concurrent.Future
   * @author Mikko Hilpinen
   * @since 17.07.2024, v1.8
   */
-case class GetEcho(content: Value = Value.empty, override val method: Method = Post, requestedStatus: Status = OK)
+case class GetEcho(content: Value = Value.empty, override val method: Method = Post, pathParams: Model = Model.empty,
+                   requestedStatus: Status = OK)
 	extends ApiRequest[Model]
 {
 	override def path: String = "echo"